rdbms: Introduce shuffle sharding to replace groups
Shuffle Sharding is rather simple: You pick a set of hosts based on an
identifier from the request (e.g. IP) and then use randomness and
weights to pick the final target.

This is made by Amazon's Route 53 and is heavily used in the industry
and is much more robust than our bespoke group system.

While this patch is adding somewhat of a complexity, once we start to
wind down groups in rdbms library, the code will become much simpler.

Also fixing some issues in tests because they were not working as
expected and noticed them during writing tests for shuffle sharding.

Bug: T405087
Change-Id: I6d21701d5d6d80059825f596081336f1b6e322ce