htmlform: Allow 'hide-if' and 'disable-if' params for no-JS users
Fields using the 'hide-if' and 'disable-if' params are always shown
and enabled for no-JavaScript users, in order to let them have full
use of the form. Sometimes, though, the forms are complex enough that
displaying all these fields is impractical.

Add 'hide-if-nojs' and 'disable-if-nojs' params that hide/disable
fields for no-JS users too. These rules are also processed when
initially rendering the form (generating the HTML server-side), so
to use the hidden/disabled fields, no-JS users must submit the form.

Bug: T320871
Change-Id: Ie4eaea5fe0b2fa4b420f3af511bb6fbf463ef8b0