tests: Assert (rev)deletion result in MediaWikiIntegrationTestCase
The `deletePage` and `revisionDelete` methods do not expose the
StatusValue object to the caller, nor do they check it themselves.
So, any failure is going to remain unnoticed and likely confuse the
test author.

Add a check that the action succeeded, so that callers can still "call
and forget", and be sure that everything is as they expect it to be.

Do not use `assertStatusGood` or similar methods to avoid increasing the
assertion count of the test.

Also remove redundant `deletePage` calls after FileDeleteFor::doDelete,
as that already deletes the page, so this redundant deletion would
always fail.

Change-Id: I86fce7c84a074bb100ad7396177d30c5da9e3bc0