RevisionStore: Correctly respect the search limit for identical revs
In Ife55590a2c983be we updated the code to no longer use rev_sha1 but to
take a look at the content_sha1 values of its slots and matching them.
But this approach actually causes the code to no longer correctly
respect the search limit as described in the method comment in case of a
revision having more than one slot.

Bug: T389026
Change-Id: I602f81cfeba0a2b2f32e36cc92a57a77999249fd