Add a warning when creating a new account while already logged in
To clearly display that this is an unusual action, add a warning
message about it, an action button (to continue with the currently
logged in account) and a form heading that once again points out that
you're creating another account.

Update the existing warning when logging in while already logged in
in a similar manner.

None of this prevents users who actually need to create additional
accounts from doing so.

Bonus changes:
* Add 'userlogin-loggedin-prompt' and 'createacct-loggedin-prompt'
  messages, empty by default, which can be customized by wiki admins
  to show messages about account switching and account creation for
  others, respectively.
* Hide the benefits block when creating another account.

Bug: T284927
Change-Id: If544c2d031a8151352925bd85e0a27d94a018df2