Make the login and signup forms wider
According to Codex Style Guide:
https://doc.wikimedia.org/codex/latest/style-guide/constructing-forms.html#layout
forms "should span the entire parent container of the page".

We need to leave some space for the account benefits column,
so we can't quite do that, but we can make it less cramped.

Allow the benefits column to shrink a little, so that this change
doesn't push it offscreen at common screen resolutions.

Bug: T355853
Bug: T407172
Change-Id: I70bdbd414e2dc871dadc58922b34acb91cd66381