recentchanges: Make categorization conflict with patrol status
Categorization entries are always marked as autopatrolled.
Therefore, when a patroller searches for unpatrolled
(and/or manually patrolled) categorization entries,
the query is guaranteed to return no results.

By registering these options as conflicting, the interface
prevents running expensive queries (that are bound to stop
when the whole table has been scanned) and provides
the patroller an explanation.

This is also motivated as an optimization for T89582.

Change-Id: Ic586982629f7c1fa3cf88df51bba2320caaf43d6