specials: Handle UploadStashException on Special:Upload
When the first upload via Special:Upload fails with a recoverable error
(like short name or invalid letters), the file is stashed (visible via
Special:UploadStash) and when the form is re-submitted, but the stashed
file is not found in the stash, a fatal error is shown.

Catch the exception and show an error message instead.
The message may not help the user to know what to do next, but it should
be better as the exception (avoids also log spam).

Not sure when the stashed file is not found and this error can happen.

Bug: T395045
Change-Id: Ibd22601a47c0b386b09e8fe6bb6cce0bd40dff4f