Make ProtectionIndicators enabled by default
For a long time protection indicators have been manually added to
sites by communities and administrators. They should be provided by
default to all new installs, to make it easier to start a wiki and to
provide transparency to its community. More advanced communitiies can
still disable or override the functionality to be more advanced or to
their liking.

Bug: T405719
Change-Id: I20c8c337db3f0c5a6273dd1da1185cd49427361f