mediawiki.util: Make addPortlet selectorHint do what the docs say
It was being inserted before the provided selector instead of after.
Based on analyzing Common.js code and gadgets that do this by hand
today, most seem to want to add something *after* a given menu.

This also makes sense from a UX and performance perspective, as a
common use case is to add a menu after 'p-tb' (Toolbox) in Vector
and Monobook. This would otherwise add the menu before the toolbox,
above the fold, which would cause mis-clicks and a layout shift.

The docs already stated this, but the test didn't assert it, and
the implementation actually inserted it before.

Change-Id: I672c4416337ca581ee255ccdac047abae501e647