updater: Proper PRIMARY KEY modification
The current addIndex methods only checks for an index of the given name
to exist. The PRIMARY KEY is is a table constraint which we want to
sometimes change.

So we will introduce a new getPrimaryKeyColumns() method to rdbms which
allows us to check which columns are currently part of the PRIMARY KEY
in order to check if it needs to be modified.

Bug: T299951
Change-Id: I46d254a1e15884c326104bd5ad9874ad54cb9f4a