session: Handle an edge-case in MultiBackendSessionStore::set()
`MultiBackendSessionStore::set()` will call `::getActiveStore()` which
will end up doing multiple store lookups (`get()`). This is not an
issue normally but causes a lot of cache misses when we're doing it
for cases where the key that is being looked up is not even in the
backend.

`SessionManager::generateSessionId()` calls `set()` on session store
with `false` as the value and a write to the cache ONLY instead of a
write to both the cache and the backend store. This needs to be handled
so early to avoid pointless cache lookups.

Why:

- Doing a lot of store lookups for a key that we know that has not been
  persisted in the backend would create a lot of cache misses which is
  going to put a lot of pressure on the infra.
- This caused the revert in Id029bf1d2c476c9039a56 (5b53bd2).

What:

- This patch handles the edge-case described above and should resolve
  the issue that caused the revert of the config to group1.
- What we do here is just writing to both store's cache directly so that
  we can always have the value if needed. Note that we're not writing to
  the actual backend in this case and no actual backend fetches will be
  performed for keys that do not exist in the backend.

Bug: T402808
Change-Id: I73672ba6aba8100e09c1b429bd6d2458f223f994
(cherry picked from commit c44e3ed7499c5e4ba2b955867a2d5ab4c21c06aa)