Support MW REST API endpoint and module deprecation
OpenAPI supports a "deprecated" property, indicating if an endpoint
is deprecated. Also, https://www.rfc-editor.org/rfc/rfc9745.txt
proposed a "Deprecation" response header informing callers of an
endpoint that it is deprecated.

Support both of these by allowing a "deprecationSettings" object
in our module definition files and flat route files.

Adding "deprecatedSettings" for an endpoint will add a "deprecated"
property to the endpoint's generated OpenAPI spec.

If the "deprecatedSettings" object contains a "since" property, we will
add a "Deprecation" header to responses for that endpoint.

Deprecating a module applies all the above to all endpoints in the module.

Bug: T405038
Change-Id: I7c2d02f9fd5bd07b4f865a680d2965a81e05eca1