selenium: Replace mwbot
Replace mwbot since it's unmaintaned and have dependencies that are
marked as security issues.

The replacement uses NodeJS built in fetch, introduced in NodeJS 18
and was stable in 21. That means NodeJS 20 users will use "unstable"
version but I don't think its a big problem. We also use some other
"new" introduced functionality as Headers, getSetCookie and AbortSignal
and all exists in 18 or later.

The actual calls to API was ported from mwbot.

Bug: T404361
Change-Id: Ibf48dc7bdd41f9e089351e4c5886c6777e195b47