Fully remove $wgDBDefaultGroup
This config variable does not make any sense to me. The biggest usecase
is when running a dump or some other environment, it sets the global
value so when ILBF reads that value, it can determine what is the
default group which also is an anti pattern since the moment ILBF is
initialized, any change to the value will be noop and as result, we have
a lot of code in quite early stages just to read the config.

This shouldn't be a config. A sysadmin of a third party mediawiki
installation shouldn't be setting the default group value.

This allows us to simplify the logic in various places.

Bug: T405087
Change-Id: I973a45815d07de83c0fa0e5259b72732240911d5