recentchanges: Migrate RecentChangesLinked to the new system
Add ChangesListQuery::requireLink() which implements the functionality
of Special:RecentChangesLinked.

A comment from 2008 complained that the awkward SQL, with a union
in a subquery, solely exists to support the assumptions of the parent
class, and suggests an emulated union like Special:Whatlinkshere. This
would indeed be more portable and would be less likely to end up in the
slow query logs. So, do that.

Bug: T404416
Change-Id: Ib305877c2e6878d348ecc547bb3be83d32174cfa