tests: Mark ExpectCallbackTrait as stable to use
CommunityConfiguration introduced a domain event,
and we would benefit from expectDomainEvent(), so that
we can assert the domain event fires when it is supposed to.

As of now, the trait is not marked as stable to use,
which means it would be against the Stable Interface Policy
to use it from CommunityConfiguration.

Bug: T405961
Change-Id: Ibd3f04f23180444bc95013574a45ecf7911e00a5