recentchanges: Migrate change tags to the new system
Add ChangeTagsCondition, which is responsible for filtering changes
lists by change tags. Move the relevant code from the three
ChangesListSpecialPage subclasses. The caller ends up being the same in
all three classes, so move it to the base class.

SpecialRecentChanges had isDenseTagFilter(). Instead of activating it
only on that one special page, introduce a density heuristic derived
from the active conditions, which is true on some RC queries but false
on Watchlist and RCL queries.

The complex interaction between ChangeTagsStore::modifyDisplayQuery()
and the ChangesListSpecialPage subclasses is evidence against the
idea. A generic query modification method doesn't have enough
information about the query it's running to correctly do the job. It's
simpler to just reimplement it. ChangesListQuery is a storage layer
class, so it's allowed to know the schema, including change_tag.

Bug: T404416
Change-Id: Icc03ecb8a0af011f0f9ee487975590394dd4913c