session: Do not set JWT cookies for anonymous users
It does not seem particularly useful. Plus, if there's some kind
of cookie conflict (especially on Wikimedia's shared authentication
domain where the JWT cookie is shared by multiple wikis), user
impact will be smaller this way (a missing JWT cookie is silently
created, while an anon JWT cookie when a user is expected is an error).

Bug: T399631
Change-Id: Ib1a3a291269f36ef9546a10f54aae1dd0eec3b51