Revision: use LOCAL instead of current wikiID when creating RevisionStore
Use the same normalization pattern used by other service factories (see
also T405922) and switch to LOCAL instead of using the full ID of the
current wiki. This is disabled in PHPUnit tests because
RevisionStoreDbTest relies on it for its "crossWiki" tests, sadly.

Also, in RevisionRenderer, print both wiki IDs when there is a mismatch,
so it's easier to understand what's going on.

Bug: T406777
Change-Id: I52be5f1ee59c242db6ed0efd2ecfdc380a565171