SECURITY: Sanitizer: disallow underscore and wide underscore in data-* attribute names
CVE-2025-67479

Also add wide underscore to the set of characters encoded in
attributes to ensure that Japenese double-underscore magic words
(which can begin with a double-wide underscore, U+FF3F) don't bypass
these protections.
Encode all underscores in attribute values so magic words that start
with double underscores but don't end with them are not replaced.

Bug: T407131
Depends-On: I380f7340f64b78dd5e8860c3848fda5a5bd68c6c
Depends-On: I5895954bca976f81a426f8648e24c9f2ee87fc10
Change-Id: I351edf35b965ac3cc9e50e47649efd88d238b2c2
Co-Authored-By: SomeRandomDeveloper <thisisnotmyname275@gmail.com>