i18n: deprecate double-underscore magic words which don't start/end with __
All "double underscore" behavior switches should actually begin and end
with either a double underscore or a double U+FF3F (a fullwidth
underscore used in Japanese).  Parsoid's tokenizer will not match a
localized behavior switch which does not start with double underscore.

In this patch localized aliases are added which start/end with double
underscore, and a few erroneous localizations were corrected which
had leading spaces or zero-width spaces.  In addition, pages which
use a localized behavior switch which does not start/end with double
underscore are added to a tracking category for manual fix up.

In a future release (I8dd522d605c2b9e5310d169d7c51bcf424089497),
aliases which do not start/end with double underscore will be removed.

Bug: T407289
Change-Id: I458e3c981c07394ab81259fadb3d68a0399e7953