API: imageinfo should take into account vector scaling
When the imageinfo api is requested to generate a thumbnail, that is
limited to a height we need to set a maximum for the width that is
'acceptable'.

In case of SVGs with an inherent size that is smaller than the
requested size, the maximum width would be set to the inherent width
of the SVG. Instead, simply limit it to the highest allowed width for
thumbnails.

Bug: T310433
Change-Id: Ia8ad1d62d34db37a3bb439cb14bfb88e2881bef3