specials: On upload try-again warning, do not stash again
When an user has a warning or a recoverable upload error, the uploaded
file is added to the stash (can be seen on Special:UploadStash) and the
user can change the description or file name and try again. If this also
fails the same file is stashed again under a new file key.
Avoid this additional stash and reuse the file key.
Use UploadBase::$mStashFile to hold the file from UploadStash and allow
reuse for this check.

Change-Id: I268caf8d97fc747091f0c97444d8e26671eaae09