specials: Start upload with stashed file from Special:UploadStash
Allow Special:Upload to load a stashed file and add a link from
Special:UploadStash to Special:Upload with the necessary query
parameters. This should be a low-effort way to allow reuse of stashed
files also with the UI and not only with the api.

Special:Upload already can show stashed files as needed for the
try-again workflow when upload warnings are shown. This is using the
same query parameters internally (UploadForm::getSourceSection sets the
hidden fields), but with a GET request. (POST request from
Special:Upload with wpSourceType=Stash&wpSessionKey= works already).
The use of UploadBase::createFromRequest creates the neccessary upload
base (UploadFromStash).

Use UploadBase::$mStashFile to hold the file from UploadStash to allow
checking if there is a stashed file, that should be shown on GET.

Skip unfinished chunked uploads (us_status != finished) to avoid broken
files to process.

Bug: T85561
Change-Id: Iac64a161ac16460fb12e31c47a7438f935afb209