specials: Update thumbnail html for file warnings on Special:Upload
With $wgUseLegacyMediaStyles = false there is no styling for the
shown image on Special:Upload.

Convert the html to use the figure tag with typeof similiar to the thumb
of images generates from wikitext.
Does not use functions like Linker::makeImageLink as that would show
also a image in case of thumb errors, that is not needed here.

The image is shown when after submit a upload warning exists,
like a duplicate file name.
There is also a live warning before submit via javascript,
that is not affected.

Bug: T386546
Change-Id: I7af9ebe5f1f5a34ad7c8e3b6a4709b34833d41a2