recentchanges: Temporary fix for incubator exception
WikimediaIncubator adds an option to the FormOptions in
onChangesListSpecialPageQuery, and then consumes it in
onSpecialRecentChangesPanel. This can fail with an exception after the
introduction of ChangesListQuery, since the query hook is not called when
an empty result set is forced, for example when there are conflicting
options.

So, call the main query hook after ChangesListQuery was executed, if it
wasn't called already.

I think WikimediaIncubator is the only affected extension. A comment in
Translate indicates that they tried to do the same thing but it didn't
work for them.

Change-Id: I64a5efd3f750be3235a3dd21092a5818a7a28b05