OutputPage: Remove UrlUtils::expand call from redirect destination
This code was was added in 2003 by SVN r2231, with a minor change in
2008 by SVN r30898.

Relative URLs were supported in most browsers in 2003, but the spec
did not recognise this as standard at the time. That was fixed in
RFC 7131 which updated the HTTP 1.1 spec to include it and validate
it going forward.

Bug: T406402
Depends-On: I4b562866d14f48256ba405e150088addb0745fba
Change-Id: Ia8366125cbef2c9e1740f0fa71818a19cedced55