recentchanges: Restore table qualifiers in change tag field expressions
The Translate extension joins another copy of this table, which causes an
SQL error "Column 'ct_tag_id' in WHERE is ambiguous". Prior to
ChangesListQuery, the fields were all qualified, so restore that
qualification.

There's no need to use ChangeTagsStore::DISPLAY_TABLE_ALIAS since the
alias name is internal to ChangesListQuery, we're not depending on
ChangeTagsStore to use this alias name.

Bug: T408040
Change-Id: If749173f395b36849d34cdf7b605b79c2cd2719b