ParserOutput: use null instead of INF for mMaxAdaptiveExpiry
The type of the $mMaxAdaptiveExpiry property was 'integer or INF'.
INF is a float, but methods setting MaxAdaptiveExpiry only took int.
Setting the type to `int|float` was confusing about the actual range
of values; using `?int` as the type and using `null` to represent "not
set" (instead of INF) yields more type-safe code.

This is a private property, so the actual values used in the
representation should not be visible outside the class.  The
serialized form is unchanged, since JSON didn't allow `INF`
values in the first place.

Change-Id: I5e9c27709bbe6b5541599de254a08d8c8fba9970