LanguageVariantConverter: fix a DOMException
DOMUtils::parseHTML( '' ) was returning a document which already
contained a `<head>` element, causing a DOMException when a second
head was added.  The rest of this code relies on the returned
document containing a `<body>` tag, so clarify for future maintainers
by including an empty <head> and <body> explicitly in the ::parseHTML
call.

Change-Id: I9b1cd9bc6b57c287d99105cee67d6f731703b61e