SECURITY: Always escape commas in mail encoded-words
CVE-2025-67478

In an RFC 2822 style header accepting multiple values (like the To:
header of an email), a comma is used to separate distinct values, unless
it is quoted or escaped. The logic MailAddress::toString() always quotes
values, unless the RFC 2047 encoded-word encoding logic in
UserMailer::quotedPrintable() was triggered. However, that encoding
logic until this patch left commas as is, letting those sneak up to the
final message unescaped.

This patch simply adds commas to the list of characters that cannot be
left as-is on encoded-word data.

Bug: T385403
Change-Id: I673d5eb82d80d789bd443db2db846cdc45962c39