Stats: have RunningTimer manage the initial label set
Why: In situations of recursion, downstream timers can overwrite
the labels of an upstream timer waiting for the downstream to finish.
This is because the metric instances themselves are singletons, but
RunningTimer instances are unique.

What: Have RunningTimer keep a set of labels as they were presented to
the TimingMetric instance prior to start().

As the timer is running, RunningTimer must also support changing the
labels prior to stop().

Bug: T406170
Change-Id: I47a32774c6e932088839ea77a3ca30d915a79411