revision: Allow RevisionStore methods to take PageReference instead of LinkTarget
For RevisionStore methods that take LinkTarget|PageIdentity, permit
them to also take a PageReference. This allows callers to be migrated
away from passing LinkTarget without the overhead of an additional query
to fetch the page ID. Passing LinkTarget was deprecated in 1.36 but
PageReference was introduced in 1.37.

Without this change, RevisionStore::getRevisionByTitle() would have
become redundant with getRevisionByPageId(). The naming of the method
implies that there should be a page_namespace/page_title condition.

Migrate the sole remaining use of PageIdentity in WatchedItemStore.

Change-Id: Ia6d987f222129849e6bdec3d933207e7dad229dc