ResourceLoader: Fold LessVarFileModule into FileModule
When modules are registered with a "class" key, they instantiate the
given class instead of the default FileModule. LessVarFileModule
extended the default FileModule class with the functionality to
embed localization message as variables in a Less stylesheet.

Now, without specifying the class, that functionality is available
in the default FileModule.

Bug: T398827
Change-Id: Ife2f8289b34f23d32e6201ba130dcc91ff3832a0