objectcache: Improve explanation of lowTTL and hotTTR and async regens
Also explain how to disable these, since it is not obvious whether to
use 0, false, null, INF, or -1.

I noticed that `ageNew` is the only one among the ones I'm documenting
that can only be disabled by -1, not by 0.

In this documentation-only change I'd like to explain this feature
through one consistent name, so I'm referring to this as the "hotTTR"
feature, with "ageNew" being a secondary option that influences that.
It seems clearer to disable this as `hotTTR: 0` than as `ageNew: -1`
(either way works, technically).

We could make `ageNew => 0` work by changing one condition from `< 0`
to `<= 0`, because `ageNew => 0` is not used in Codesearch Everywhere.
But, I think `ageNew => 0` has no obvious meaning by itself, I'd much
rather encourage `hotTTR => 0` instead. It seems reasonable to suggest
that a has a "new" threshold of 0s, meaning it is always new. It
does not seem reasonable to suggeset that a key must be refreshed
at a rate of 1x every 0s. And, using the hotTTR option tells a more
consistent story between documentation and call sites.

Bug: T353180
Change-Id: I6c9a7eb0294404be2ddf0a887db9fb1b9a249d95