objectcache: Fix flaky testPreemptiveRefresh
The test for the lowTTL option (covering worthRefreshExpiring)
should disable random influence from the hotTTR option
(covering worthRefreshPopular) and vice versa.

In particular, the problem was the before last one:

```
$key = wfRandomString();
$opts = [ 'hotTTR' => 900 ];
$v = $cache->getWithSetCallback( $key, 60, $func, $opts );
$this->assertSame( $value, $v, "Value returned" );
$this->assertSame( 1, $wasSet, "Value calculated" );
```

This still had the default lowTTL=60s enabled, and that started
right away with a chance of 0.0625 (6.25%) leading to an extra
async regen 1:16 times. That is expected and not a problem in
production, but messes with the strict call count assertions
in this particular test where we want to test each of the random
factors separately.

Bug: T353180
Change-Id: I100a345429c9fe97a6a303d70157e5b33682b4a5