user: Update user_touched on db on login if it hasn't been updated
This used to happen on every login action until 2015 but it was removed
due to amount of writes it was adding to the database. But we need this
information for various reasons and a standard way to avoid too many
writes in these cases is debouncing.

It has three two periods:
 - If the write has been done too recently (here 5 minutes), just skip the write.
   - Noting that there is a jitter of 20 seconds to avoid stampede.
 - If it's more than that, update the value.

This makes sure rouge bots that sometimes login thousands of times every
minute won't overwhelm the system with writes.

It basically reverts I96b30f5c9c8 but adds back the functionality in a
robust way.

Bug: T406724
Change-Id: Ida24e4500a23ee054399a5d7c61a848f9fbd0c67