RestrictionStore: Switch order between pr_cascade and links queries
We originally queried page_restrictions first since we expected there to
be lass than 100 pages with pr_cascade=1.

But we actually missed metawiki where all pages related to CentralNotice
are automatically cascade protected and there are ~50.000 in total.

So we need to swich this around to release pressure from the dbs.

Bug: T408801
Change-Id: I4eba736a07cf1f0b5edf12c8a30898ce27a9ea73