Mend handling of revision links on Special:Undelete
When trying to modify visibility for deleted file revisions
from Special:Undelete, the original redirectToRevDel() was
missing additional functionality to handle mixed or nothing
selections.

Also, the generated URL was lacking proper form and ID handling,
which is now included. Additional error messages included as well
for better UX and understanding of the behavior.

Bug: T329670
Bug: T379070
Change-Id: I90d956fd413244921699d8e25e379e2821ee350d