BacklinkCache: Switch order between pr_cascade and links queries
We originally queried page_restrictions first since we expected there to
be lass than 100 pages with pr_cascade=1.

But we actually missed metawiki where all pages related to CentralNotice
are automatically cascade protected and there are ~50.000 in total.

So we need to swich this around to release pressure from the dbs.

Bug: T408801
Change-Id: I0b90b7f78221f0407e028340a2e0fc6695963e1f