recentchanges: Fix highlights where more than one action is defined
Some highlights have an inverted sense compared to the filters they are
attached to. I introduced the "highlight" property in the definition to
account for this. However, I neglected to update an instance of $action,
so for the "last" and "previous" filters, the sense of the highlight was
accidentally inverted.

Bug: T409020
Change-Id: I59ef6d99792b58a6b854c9e4e3b459fa48afa81c
(cherry picked from commit d9f65865cb9c0cbd45caa8fad8e476ee04e39150)