Add some type declarations to database domain-related code
Use PHP 8 union types such as `string|false` to enforce the expected
types. This guarantees early failures in case of mismatches, whereas
previously the invalid value would be passed around a lot before ending
up in some place that didn't support it, as seen in T409088.

Also add a return type declaration to `User::getWikiId` so that PHPUnit
can automatically set a valid return value when none is specified on a
User mock.

Note that adding type declarations to parameters is a
backwards-compatible change, because of parameter type contravariance.
Return type declarations are added to methods that are not possible or
meant to be redeclared in subclasses.

Drop doc comments with type declarations that are no longer necessary.

Bug: T409088
Change-Id: I71c0282a5c59dd66f48ccfe4bf8d2a89bb321e1a