watchlist: Remove WatchedItemQueryService::getWatchedItemsWithRecentChangeInfo
In 1.45 I rewrote ApiQueryWatchlist so that it doesn't call this method.
That left no known callers, since it was very specific to the API.
Now, remove the method and hard-deprecate the hooks it called.

I'm not removing the hooks just yet since it is still possible for
extensions to support multiple versions of MediaWiki by handling both
sets of hooks, as we did for ORES.

Bug: T407087
Change-Id: Ic1d58b4c1d923c893fd181ce7ef29ce5e933e8d8
Depends-On: Iadba9b576851f36838d62a504bc66760077fa29c