recentchanges: Fix watchlistactivity=all, i.e. seen/unseen conflict
The "seen" and "unseen" filters are supposed to implement a union, so
that if both are selected, all changes are shown. This was broken by
2a60c670cd4aa7c69a36195dbb4aa26b978737f3

So, make SeenCondition take a boolean value so that both false and true
can be required. The implicit union between required values means that
all results are shown.

Review of other cases of require and exclude being used in the same
filter group:

* hidemajor/hideminor: by the names, an empty result set is correct.
  If both are selected, the JS frontend doesn't send either. The legacy
  UI doesn't have hidemajor. So it doesn't come up in practice.
* hidebots/hidehumans: same
* hidemyself/hidebyothers: same
* hideliu/hideanons: In this case, both can be selected in the legacy
  UI. The conflict is detected and the special page redirects to disable
  both options.

The bug is specific to require/exclude conflict with a string options
filter with isFullCoverage, and seen/unseen is the only instance of
that.

With "hide" names and full coverage, selecting all options should
naturally lead to an empty result set, hence the existing use of
require/exclude which provides this behaviour.

With positive names like "seen" and "unseen", a union is expected and
thus require should be used consistently.

Bug: T408167
Change-Id: I9814af27105963213a8ea5ced10d118906353a62