REST: enable deep linking in swagger ui for the REST Sandbox
Swagger UI allows "deep linking" to specific endpoints, if the
"deepLinking" config value is set. This allows constructing links
that scroll directly to a specified endpoint and automatically
expand that endpoint. This is useful for API documentation.
We did not previously set the necessary swagger config value.
Enable it, so that we can use deep linking in our docs.

Bug: T405972
Bug: T409198
Change-Id: I86c089626b096563eead028aff1fef15aa51b268