maintenance/getConfiguration.php: Fix null warning and serialize error
When invoked without options, the default has two issues:

```
PHP Deprecated: strtolower(): Passing null to parameter of type string
is deprecated in /mediawiki/maintenance/getConfiguration.php on line 135
```

This is because despite the constructor and validateParamsAndArgs
treating --format as optional with default, execute() did not.

```
Exception: Serialization of Closure is not allowed
```

... from $wgHooks

```
LogicException: Instances of OutputPage are not serializable
```

... directly from $wgOut, and indirectly via ContextSource
from $wgUser and $wgRequest.

Change-Id: I2efa7835ff47084472c7fb079f93077b736057cb