TemplatesOnThisPageFormatter: Remove custom protection level labels
Instead of having two separate messages for the two built-in
protection levels:

  "template-protected": "(protected)"
  "template-semiprotected": "(semi-protected)"

Use the same messages that are used to display protection levels
elsewhere in the interface, wrapped in the "parentheses" message:

  "restriction-level-sysop": "fully protected"
  "restriction-level-autoconfirmed": "semi protected"

This pattern was already used for any custom protection levels.

The old messages can still be used if they are customized on-wiki,
but they are now disabled by default.

Change-Id: Ic406b680b7610d6110d2b7876585733b7641c2fb