Add table to Special:WatchlistLabels
Add a table to Special:WatchlistLabels that lists all of a user's
current labels. A 'new label' link is in the table header,
and it showns (on the /edit subpage) an form for creating a new
label. No label editing is added at this point (as this patch is
probably already doing more that was supposed to).

Validation is added to the form to constrain the label length.

Bug: T408775
Change-Id: Icd1bde77bb100a4fb0d5f7e5a901d2f08a282f77