RevisionRecord: Stop using rev_sha1
There is a strong warning for the computation being potentially slow.
I checked the code and it seems that everywhere where the slots are
loaded, the comment table is being joined and thus content_sha1 is
available. So this is only slow if the content_sha1 field is not
populated which is no longer the case anywhere after fixing dewikivoyage
and itwikivoyage.

Bug: T389026
Change-Id: Ic986570718d93df652fcda8df67657141626bef1