ApiQueryInfo: Switch order between pr_cascade and templatelinks query
We originally queried page_restrictions first since we expected there to
be lass than 100 pages with pr_cascade=1.

But we actually missed metawiki where all pages related to CentralNotice
are automatically cascade protected and there are ~50.000 in total.

So we need to swich this around to release pressure from the dbs.

Bug: T408801
Change-Id: I21621f4c4b74a472f83e7bbd59478568f24e9145