Mark top-level page title with appropriate page language and direction
This makes a big difference when the user's interface language
differs in directionality from the page content language.

Note that the outer <h1> still has lang/dir matching the user
interface language, so the overall heading alignment is consistent
with the skin; a separate <span> wrapper around article titles
uses the page content language (which may be different from the
wiki content language).

Bug: T36514
Change-Id: I846fde4e8ac05a6e9d387641e14d44189bb332ff