HTMLForm: Unify option for excluding special pages in title suggestion
The PHP HTMLTitleTextField, has an option for validating title is creatable (not special page) since initial commit. But this option
was not forwarded to the JS implementation, so it only works after
form submission.

In T208355 (a7fd92a95e), this validation after submission was not acceptable, so new option 'excludeDynamicNamespaces' was added to
the JS widget for use in HTMLTitlesMultiselectField.

This patch unifies the logic by forwarding the older option to the widget.

The 'excludeDynamicNamespaces' option has been removed without deprecation as there are only 3 public usages which will be fixed together with this patch.

Also fix the option in SpecialBlock.php, the only usage in core.

Bug: T409700
Change-Id: I81db7f1c9f7fd4449529098e9a5b78e79d1ef9a6