T406897: Add aria-labelledby to <section> tags to improve accessibility
`<section>` is assigned a "generic" ARIA role without this attribute,
i.e. it's semantically meaningless. With the presence of
aria-labelledby, however, it's being assigned a "region" ARIA role,
which is a landmark role and denotes "document areas the author deems
significant", with all benefits of a landmark role (more semantics for
the page, more navigation for screen reader users, etc).

This patch adds the attribute in the postprocessing pipeline used
for read views, so it is not present in "canonical" HTML used by
clients and VisualEditor.

Bug: T406897
Change-Id: I3433cb6279647ca504cc87de7dcc5fa38d86688c