Use first class callable syntax in more places
Convert array and string callables to PHP 8.1 first class callable
syntax, excluding built-in functions (e.g. leaving `'strlen'` unchanged,
instead of changing it to `strlen( ... )`).

These changes were made by `phan --automatic-fix` with the new plugin:
https://gerrit.wikimedia.org/r/c/mediawiki/tools/phan/+/1160388

Fixing a few of the issues caused test failures or other Phan issues
to be detected, and I undid the fixes. These issues should be suppressed
when upgrading to the new Phan config with the plugin later.
Two of them were genuine problems in the code or docs, fixed in:
* Ib83cde703120d1c09555edb37008acf6d3a64277
* I05bdcb01359b245ef8378f3e30bbda8922f0dbe8

During code review we discovered that on PHP versions before 8.1.4,
first class callable references to user-defined (not built-in)
functions (not methods) may fail with nonsensical ArgumentCountError.
I kept those as strings and added comments noting the issue.

Bug: T397107
Change-Id: I830a15fa61b0ee362844fc1a122613abdb2ac738