Api: Avoid re-stashing on publish with warnings via action=upload
A publish request to the upload api does not contains a file,
because the file from stash is the file to publish.

When the publish step founds warnings, like duplicate name,
the api is stashing the file again, but that is not needed, because the
stashed file still exists for a re-try.

In case of async publish the api result contains a
"UploadStashBadPathException" stash error, because the stashed file is
not loaded in the request (it loaded from jobqueue later), but that is
not a real error.

Check if the UploadFrom*-class has already an stashed file and not try
to stash again. Since d6daeee1a0 a similiar check is done on
Special:Upload to avoid re-stashing files on warnings.

Change-Id: I0c5a7188179eb110e7e9c612e6cf3691ce95aae6