title: Disallow string as namespace index in NamespaceInfo service
I'm not sure if we should mark this as a breaking change. It doesn't
look like this feature was ever documented anywhere. It doesn't
belong in this class, in my opinion. Basic validation of user input
should be done before any of these methods are called.

This feature was added in 2020 via I66ea07f (see T253098). Shouldn't
we use PHP's language-level type declarations to do this now?

Bug: T253098
Depends-On: Ifb8240800b4c111e46f8e29a0d40aa64d49207a1
Change-Id: I06dd21c0777e3cb8d0dccdee92e4a790a0017680