Post-process the ParserOutput in Article closer to the cache access
This first patch replaces the use of OutputPage::addParserOutput in
Article by post-processing the ParserOutput first, and then adding it to
the OutputPage via the new method
OutputPage::addPostProcessedParserOutput. The goal of this change is to
make the cache access and the OutputTransform pipeline closer to one
another, so that we can eventually get a processed ParserOutput directly
from the cache.

Bug: T348255
Change-Id: I35c7afffb6a5cc0f50722a67774f2d7dffbe9b5f