mediawiki.jqueryMsg: Support self-closing HTML tags
This allows <br>, <wbr> and <hr> to be used in messages that are
parsed by jqueryMsg.

After parsing the opening tag, we check if the tag is in the
allowlist for self-closing tags. If it is, and its attributes are
valid, we return it as an HTML element without trying to parse an
end tag. If the attributes are not valid, we escape it.

While there are some tags like <li> or <dd> that can be closed
either with a normal closing tag or by self-closing, this patch
doesn't allow them for now to keep the implementation simple.

Bug: T295568
Change-Id: I4db9a4cd7c81e096220023d3c806781a9e03adfa
(cherry picked from commit 2bcc7bb3cf6fd2a7426f7d98cb5b654b635f66eb)