Add MWTimestamp::fromMW()
Add an efficient constructor to create an MWTimestamp from a string
in the TS_MW format.  Ideally we'd have an even thinner wrapper
around the string that would defer actual conversion until needed,
but this is at least better than the full ConvertibleTimestamp
constructor which tests against a dozen different time string
formats.

Change-Id: Ib242c555dd898fe07d3fd97a8a7c5b94ec6ddfd6