OutputTransformPipeline: deprecate ::run() without ParserOptions
All callers which passed `null` should have been removed with the
deprecation (in MW 1.44) and removal (in MW 1.45) of
ParserOutput::getText().

As such it should be safe to deprecate a `null` ParserOptions
passed to OutputTransformPipeline::run().

The intent is to backport this deprecation to MW 1.45 and so
the deprecation notice has been added to RELEASE-NOTES-1.45.
This will let us actually remove the nullable type in MW 1.46.

Codesearch:

https://codesearch.wmcloud.org/deployed/?q=getDefaultOutputPipeline

(it is a little difficult to search just for calls to `run`,
but all calls should follow ::getDefaultOutputPipeline).

Change-Id: I27d821c95be14ee5acdff75830a32e61caf4b8f1